♻️ Move shared test data into tests/helpers.py
